fix(web): split segmented control variants
